Added new manager specifically for handling LCA
