+++ src/ConflictAncestorResolverInterface.php
