+++ src/ConflictManager.php
@@ -8,15 +8,6 @@ class ConflictManager
 {
 
     protected $resolvers = [];
-
-    /**
-     * 
-     * @param ConflictAncestorResolverInterface $resolver
-     */
-    public function addLcaResolver(ConflictAncestorResolverInterface $resolver)
-    {
-        $this->resolvers[] = $resolver;
-    }
     
     public function addConflictResolver(ConflictResolverInterface $resolver)
     {
@@ -28,21 +19,6 @@ class ConflictManager
         foreach ($this->resolvers as $resolver) {
             if ($resolver->applies()) {
                 return $resolver->SimpleMergeResolver($revision1, $revision2, $revision3);
-            }
-        }
-    }
-
-    /**
-     * @param RevisionableInterface $revision1
-     * @param RevisionableInterface $revision2
-     * 
-     * @return int revision_ID
-     */
-    public function resolveLowestCommonAncestor(RevisionableInterface $revision1, RevisionableInterface $revision2)
-    {
-        foreach ($this->resolvers as $resolver) {
-            if ($resolver->applies()) {
-                return $resolver->resolve($revision1, $revision2);
             }
         }
     }
